+++ fast_rx/lib/src/rx/rx.dart
@@ -1,24 +1,18 @@
 import 'dart:async';
 import 'package:fast_rx/fast_rx.dart';
-import 'package:fast_rx/src/rx/rx_zone.dart';
-import 'package:fast_rx/src/rx_notifier.dart';
-import 'package:flutter/foundation.dart';
 
 /// Base class for reactives
 abstract class Rx<T> {
-  bool get _zoned => Zone.current[zonedKey] ?? false;
-
   final StreamController<T> _controller = StreamController.broadcast();
 
   /// Stream of value changes
-  @nonVirtual
   Stream<T> get stream => _controller.stream;
 
   /// Register with the [RxNotifier] for UI updates
-  @protected
+  ///
+  /// Should not be called directly
   void register() {
-    if (_zoned) return;
-    RxNotifier.instance.addStream(stream);
+    RxZone.current.registrar?.call(stream);
   }
 
   /// Run [action] with registration and notifications disabled. Will notify
@@ -26,8 +20,7 @@ abstract class Rx<T> {
   /// implementation decides a notification should be sent.
   ///
   /// Returns true if a notification was attempted during [action]
-  @protected
-  bool run(VoidCallback action, {bool notify = true}) {
+  bool run(RxAction action, {bool notify = true}) {
     var notified = false;
     runRxZoned(
       () {
@@ -50,11 +43,12 @@ abstract class Rx<T> {
   void notify();
 
   /// Notify listeners with the given value
-  @protected
-  @nonVirtual
+  ///
+  /// Should not be called directly
   void notifyWithValue(T value) {
-    if (_zoned) {
-      Zone.current[notifierKey].call(identityHashCode(this));
+    final notifier = RxZone.current.notifier;
+    if (notifier != null) {
+      notifier(identityHashCode(this));
       return;
     }
     _controller.add(value);
@@ -82,8 +76,7 @@ class RxValue<T> extends Rx<T> {
   /// Used to prevent unnecessary calls to [register] in internal methods such
   /// as [copyValue], [shouldNotify], or methods that do not return a value
   ///
-  /// Overrides should be annotated with `@protected`
-  @protected
+  /// Should not be called directly
   T get unregisteredValue => _value;
 
   /// Set the current value
@@ -148,10 +141,8 @@ abstract class RxObject<T> extends RxValue<T> {
   /// Should only be used in methods that return a value.
   /// Otherwise, use [unregisteredValue].
   ///
-  /// Overrides should be annotated with `@protected`
+  /// Should not be called directly
   @override
-  @protected
-  @mustCallSuper
   T get value {
     return super.value;
   }
@@ -159,8 +150,6 @@ abstract class RxObject<T> extends RxValue<T> {
   // coverage:ignore-start
   /// Unused for RxObject
   @override
-  @protected
-  @nonVirtual
   set value(T value) {
     throw RxObjectValueIsReadOnly();
   }
@@ -172,14 +161,14 @@ abstract class RxObject<T> extends RxValue<T> {
   ///
   /// Example: Used by fast_rx_persistence to set [_value] from a key/value
   /// store after initialization
-  @protected
-  @nonVirtual
+  ///
+  /// Should not be called directly
   void internalSetValue(T value) {
     _value = value;
   }
 
   @override
-  bool run(VoidCallback action, {bool notify = true}) {
+  bool run(RxAction action, {bool notify = true}) {
     final bool notified;
     if (notify) {
       notified = notifyIfChanged(() => super.run(action, notify: notify));
@@ -197,7 +186,8 @@ abstract class RxObject<T> extends RxValue<T> {
   /// Notify if [transform] changed the value such that [shouldNotify] returns true
   ///
   /// If the value is guaranteed to change, use [notify] instead
-  @protected
+  ///
+  /// Should not be called directly
   U notifyIfChanged<U>(U Function() transform) {
     final old = copyValue();
     final result = transform();
@@ -209,15 +199,11 @@ abstract class RxObject<T> extends RxValue<T> {
 
   /// Copy the value for update emission
   ///
-  /// Overrides should be annotated with `@protected`
-  @protected
-  @visibleForTesting
+  /// Should not be called directly
   T copyValue();
 
   /// Check if the value has changed
   ///
-  /// Overrides should be annotated with `@protected`
-  @protected
-  @visibleForTesting
+  /// Should not be called directly
   bool shouldNotify(T oldValue);
 }