+++ fast_rx/test/rx/composite/rx_composite_list_test.dart
@@ -1,6 +1,6 @@
 import 'package:fast_rx/fast_rx.dart';
 import 'package:fast_rx_test/fast_rx_test.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
 void main() {
   final shouldNotify = <RxTest<RxCompositeList<RxInt>>>[
@@ -212,7 +212,7 @@ void main() {
   ];
 
   test('RxCompositeList notifications', () {
-    expectRxNotification(
+    expectRx(
       shouldNotify: shouldNotify,
       shouldNotNotify: shouldNotNotify,
     );