+++ fast_rx/test/rx/object/rx_list_test.dart
@@ -1,6 +1,6 @@
 import 'package:fast_rx/fast_rx.dart';
 import 'package:fast_rx_test/fast_rx_test.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
 void main() {
   final shouldNotify = <RxTest<RxList<int>>>[
@@ -50,17 +50,12 @@ void main() {
     RxTest(() => [0, 1, 2].rx, (rx) => rx.shouldNotify([0, 1, 2, 3])),
   ];
 
-  test('RxList notifications', () {
-    expectRxNotification(
-      shouldNotify: shouldNotify + shouldNotifyAndRegister,
-      shouldNotNotify: shouldRegister + shouldNotNotifyOrRegister,
-    );
-  });
-
-  test('RxList registration', () {
-    expectRxRegistration(
+  test('RxList registration and notifications', () {
+    expectRx(
       shouldRegister: shouldRegister + shouldNotifyAndRegister,
       shouldNotRegister: shouldNotify + shouldNotNotifyOrRegister,
+      shouldNotify: shouldNotify + shouldNotifyAndRegister,
+      shouldNotNotify: shouldRegister + shouldNotNotifyOrRegister,
     );
   });
 }