+++ fast_rx/test/rx/rx_object_test.dart
@@ -1,5 +1,5 @@
 import 'package:fast_rx_test/fast_rx_test.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'rx_tuple.dart';
 
 void main() {
@@ -25,17 +25,12 @@ void main() {
     RxTest(() => Tuple(1, 2).rx, (rx) => rx.internalSetValue(Tuple(1, 3))),
   ];
 
-  test('RxObject notifications', () {
-    expectRxNotification(
-      shouldNotify: shouldNotify,
-      shouldNotNotify: shouldRegister + shouldNotNotifyOrRegister,
-    );
-  });
-
-  test('RxObject registration', () {
-    expectRxRegistration(
+  test('RxObject registration and notifications', () {
+    expectRx(
       shouldRegister: shouldRegister,
       shouldNotRegister: shouldNotify + shouldNotNotifyOrRegister,
+      shouldNotify: shouldNotify,
+      shouldNotNotify: shouldRegister + shouldNotNotifyOrRegister,
     );
   });
 }