+++ fast_rx/test/rx/rx_value_test.dart
@@ -1,10 +1,10 @@
 import 'package:fast_rx/fast_rx.dart';
 import 'package:fast_rx_test/fast_rx_test.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
 void main() {
   test('RxValue registration', () {
-    expectRxRegistration<RxString>(
+    expectRx<RxString>(
       shouldRegister: [
         RxTest(() => ''.rx, (rx) => rx.value),
       ],